+++ neoscore/interface/qt/viewport.py
@@ -1,7 +1,11 @@
 from PyQt5 import QtGui, QtWidgets
 
+from neoscore.core.key_event import KeyEventType
 from neoscore.core.mouse_event import MouseEventType
-from neoscore.interface.qt.converters import q_mouse_event_to_mouse_event
+from neoscore.interface.qt.converters import (
+    q_key_event_to_key_event,
+    q_mouse_event_to_mouse_event,
+)
 
 _NO_DRAG = 0
 _SCROLL_HAND_DRAG = 1
@@ -27,24 +31,24 @@ class Viewport(QtWidgets.QGraphicsView):
         # manually in the main window refresh function.
         self.set_auto_interaction(True)
         self.setViewportUpdateMode(_NO_VIEWPORT_UPDATE)  # noqa
+        self.mouse_event_handler = None
+        self.key_event_handler = None
 
     def set_auto_interaction(self, enabled: bool):
         """Set whether mouse and scrollbar interaction is enabled."""
-        self.mouse_interaction_enabled = enabled
+        self.auto_interaction_enabled = enabled
         if enabled:
             self.setDragMode(_SCROLL_HAND_DRAG)  # noqa
             self.setHorizontalScrollBarPolicy(_SCROLL_BAR_AS_NEEDED)  # noqa
             self.setVerticalScrollBarPolicy(_SCROLL_BAR_AS_NEEDED)  # noqa
-            self.setFocusPolicy(_FOCUS_POLICY_STRONG_FOCUS)  # noqa
         else:
             self.setDragMode(_NO_DRAG)  # noqa
             self.setHorizontalScrollBarPolicy(_SCROLL_BAR_ALWAYS_OFF)  # noqa
             self.setVerticalScrollBarPolicy(_SCROLL_BAR_ALWAYS_OFF)  # noqa
-            self.setFocusPolicy(_FOCUS_POLICY_NO_FOCUS)  # noqa
 
     def wheelEvent(self, event):
         """Implementation of Qt event hook for zooming with the mouse wheel."""
-        if not self.mouse_interaction_enabled:
+        if not self.auto_interaction_enabled:
             return
         zoom_in_factor = 1.25
         zoom_out_factor = 1 / zoom_in_factor
@@ -68,30 +72,50 @@ class Viewport(QtWidgets.QGraphicsView):
         super().scrollContentsBy(*args)
         self.viewport().update()
 
+    # Input event handler overrides
+
     def mouseMoveEvent(self, e):
         if self.mouse_event_handler:
             self.mouse_event_handler(
                 q_mouse_event_to_mouse_event(e, MouseEventType.MOVE)
             )
-        super().mouseMoveEvent(e)
+        if self.auto_interaction_enabled:
+            super().mouseMoveEvent(e)
 
     def mousePressEvent(self, e):
         if self.mouse_event_handler:
             self.mouse_event_handler(
                 q_mouse_event_to_mouse_event(e, MouseEventType.PRESS)
             )
-        super().mousePressEvent(e)
+        if self.auto_interaction_enabled:
+            super().mousePressEvent(e)
 
     def mouseReleaseEvent(self, e):
         if self.mouse_event_handler:
             self.mouse_event_handler(
                 q_mouse_event_to_mouse_event(e, MouseEventType.RELEASE)
             )
-        super().mouseReleaseEvent(e)
+        if self.auto_interaction_enabled:
+            super().mouseReleaseEvent(e)
 
     def mouseDoubleClickEvent(self, e):
         if self.mouse_event_handler:
             self.mouse_event_handler(
                 q_mouse_event_to_mouse_event(e, MouseEventType.DOUBLE_CLICK)
             )
-        super().mouseDoubleClickEvent(e)
+        if self.auto_interaction_enabled:
+            super().mouseDoubleClickEvent(e)
+
+    def keyPressEvent(self, e):
+        if self.key_event_handler:
+            self.key_event_handler(q_key_event_to_key_event(e, KeyEventType.PRESS))
+        if self.auto_interaction_enabled:
+            super().keyPressEvent(e)
+
+    def keyReleaseEvent(self, e):
+        if self.key_event_handler:
+            self.key_event_handler(q_key_event_to_key_event(e, KeyEventType.RELEASE))
+        if self.auto_interaction_enabled:
+            super().keyPressEvent(e)
+
+    # End of input event handler overrides